+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 tilestache CHANGELOG
 ===================
 
+0.15.0
+------
+- default to tornado rather than sync
+- clean up specs
+- clean up gunicorn recipe
+
 0.14.1
 ------
 - also install python-pylibmc, as there are outstanding pull requests to support it in tilestache